image2salience/requirements_http.txt
Flask==1.0.2
numpy==1.15.4
scipy==1.1.0
Pillow==5.3.0
gunicorn==19.9.0